add deepl and bing translation for yousuu reviews

+++ src/ysapp/models/_base.cr
@@ -24,9 +24,15 @@ module YS::Helpers
     end
   end
 
-  def self.read_zip(zip_path : String, file_name : String)
+  def self.read_zip(zip_path : String, filename : String)
+    read_zip?(zip_path, filename) || yield
+  end
+
+  def self.read_zip?(zip_path : String, filename : String)
+    return unless File.exists?(zip_path)
     Compress::Zip::File.open(zip_path) do |zip|
-      zip[file_name]?.try(&.open(&.gets_to_end)) || yield
+      return unless entry = zip[filename]?
+      entry.open(&.gets_to_end)
     end
   end
